backend/requirements.txt
fastapi[standard]==0.115.12
psycopg2-binary==2.9.10
